sample-01/requirements.txt
crewai #==0.11.0
python-dotenv
langchain-openai #==0.2.1